chore(tests): move tests next to actual file
