+++ src/services/data_dir.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach, vi } from "vitest";
 
-import type { getTriliumDataDir as getTriliumDataDirType, getDataDirs as getDataDirsType, getPlatformAppDataDir as getPlatformAppDataDirType } from "../src/services/data_dir";
+import type { getTriliumDataDir as getTriliumDataDirType, getDataDirs as getDataDirsType, getPlatformAppDataDir as getPlatformAppDataDirType } from "./data_dir.js";
 
 describe("data_dir.ts unit tests", async () => {
     let getTriliumDataDir: typeof getTriliumDataDirType;
@@ -42,9 +42,9 @@ describe("data_dir.ts unit tests", async () => {
     });
 
     // import function to test now, after creating the mocks
-    ({ getTriliumDataDir } = await import("../src/services/data_dir.ts"));
-    ({ getPlatformAppDataDir } = await import("../src/services/data_dir.ts"));
-    ({ getDataDirs } = await import("../src/services/data_dir.ts"));
+    ({ getTriliumDataDir } = await import("./data_dir.js"));
+    ({ getPlatformAppDataDir } = await import("./data_dir.js"));
+    ({ getDataDirs } = await import("./data_dir.js"));
 
     // helper to reset call counts
     const resetAllMocks = () => {
