+++ src/services/search/services/handle_parens.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
-import handleParens from "../../src/services/search/services/handle_parens.js";
-import type { TokenStructure } from "../../src/services/search/services/types.js";
+import handleParens from "./handle_parens.js";
+import type { TokenStructure } from "./types.js";
 
 describe("Parens handler", () => {
     it("handles parens", () => {
